requirements.txt
django>=3.0.7
django-cors-headers==3.1.1
djangorestframework==3.11.0
psycopg2-binary==2.8.5
python-decouple==3.3
dj-database-url==0.5.0
django-taggit==1.3.0
redis==3.5.1
